% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_heatmap.R
\name{epi_annotate_df}
\alias{epi_annotate_df}
\title{Title adds anotation columns to dataframe}
\usage{
epi_annotate_df(
  feature_df,
  epi_directory = system.file("exdata", package = "GenomicHeatMap")
)
}
\arguments{
\item{feature_df}{dataframe with real and random sites}

\item{epi_directory}{directory with .rds of the tracks}
}
\value{
tum
}
\description{
Title adds anotation columns to dataframe
}
