% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_random_human_position.R
\name{get_random_human_positions}
\alias{get_random_human_positions}
\title{Title}
\usage{
get_random_human_positions(nn = 10)
}
\arguments{
\item{nn}{number of randon site needed}
}
\value{
a tibble with the sites
}
\description{
Title
}
