% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_random_human_position.R
\name{get_random_positions}
\alias{get_random_positions}
\title{Get a random position in an specified genome genome}
\usage{
get_random_positions(nn = 10, r_seed = as.numeric(Sys.time()), genome = "hg38")
}
\arguments{
\item{nn}{number of randon site needed}

\item{r_seed}{seed to use for generation of random control insertion sites}

\item{genome}{genome to get random sites from. one of c('hg38','macFas5')}
}
\value{
a tibble with the sites
}
\description{
Get a random position in an specified genome genome
}
