% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCcontent.R
\name{getGCpercentage}
\alias{getGCpercentage}
\title{calulate GC for several window sizes}
\usage{
getGCpercentage(sites, column_prefix, window_size, reference_genome_sequence)
}
\arguments{
\item{reference_genome_sequence}{actual reference genome (from BSgenome)
to get reference_genome_sequence use get_reference_genome() function from
random_site.R}
}
\description{
calulate GC for several window sizes
}
\note{
start up time to load human genome is about 10 second
}
