% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_epi_df.R
\name{intsite_to_heatmap_df}
\alias{intsite_to_heatmap_df}
\title{Format}
\usage{
intsite_to_heatmap_df(intSites, group_variable = "patient")
}
\arguments{
\item{intSites}{insertion site dataframe}

\item{group_variable}{variable that will be use to group the data, each value corresponds to a
column in the final heatmap}
}
\value{
naive declaration
}
\description{
Format
}
\examples{
x <- epi_heatmap()
}
