% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_heatmap.R
\name{make_heatmap}
\alias{make_heatmap}
\title{makes heatmap figures}
\usage{
make_heatmap(roc.res, title = "heatmap")
}
\arguments{
\item{roc.res}{roc object}

\item{title}{title to display on the figure}
}
\value{
a ggplot
}
\description{
makes heatmap figures
}
