% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_heatmap.R
\name{make_roc}
\alias{make_roc}
\title{Title makes roc object from dataframe}
\usage{
make_roc(annotate_df)
}
\arguments{
\item{annotate_df}{a dataframe}
}
\value{
ROC object
}
\description{
Title makes roc object from dataframe
}
