% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_intSites}
\alias{sample_intSites}
\title{720 insertion sites}
\format{
A GRanges object
}
\usage{
sample_intSites
}
\description{
A dataset containing Granges for 720 insertion sites.
}
\keyword{datasets}
