% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_heatmap.R
\name{epi_annotate_df2}
\alias{epi_annotate_df2}
\title{Title adds anotation columns to dataframe}
\usage{
epi_annotate_df2(ff_df)
}
\arguments{
\item{ff_df}{datafram with real and random sites}
}
\value{
tum
}
\description{
Title adds anotation columns to dataframe
}
