% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_annotate.R
\name{gen_annotate_df}
\alias{gen_annotate_df}
\title{add genomic annotations}
\usage{
gen_annotate_df(ff_df, features = c("GC_content"))
}
\arguments{
\item{ff_df}{a genomic ranges}
}
\value{
the same genomic ranges object with genomic features added as extra columns
}
\description{
add genomic annotations
}
