% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_epi_df.R
\name{intsite_to_heatmap_df}
\alias{intsite_to_heatmap_df}
\title{Format}
\usage{
intsite_to_heatmap_df(
  intSites,
  group_variable = "patient",
  r_seed = as.numeric(Sys.time()),
  genome = "hg38"
)
}
\arguments{
\item{intSites}{insertion site dataframe}

\item{group_variable}{variable that will be used to group the data, each value corresponds to a
column in the final heatmap}

\item{r_seed}{seed to use for generation of random control insertion sites}

\item{genome}{genome to get random sites from. one of c('hg38','macFas5')}
}
\value{
naive declaration
}
\description{
Format
}
