% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Title}
\usage{
plot_heatmap(roc_df, title = "heatmap", star = FALSE)
}
\arguments{
\item{roc_df}{A dataframe generated from make_heatmap}

\item{title}{A title for the plot}

\item{star}{Whether to add pvalue significance as stars.}
}
\value{
a ggplot object
}
\description{
Title
}
